requirements.txt
blinker==1.8.2
click==8.1.7
filelock==3.14.0
flask==3.0.3
flask-uploads==0.2.1
flask-wtf==1.2.1
fsspec==2024.6.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.3.0
sympy==1.12.1
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.12.1
Werkzeug==3.0.3
wtforms==3.1.2
